pyproject.toml
[tool.poetry]
name = "upload-service"
version = "0.1.0"
description = ""
authors = ["Andrey \"MOU\" Larionov <anlarionov@gmail.com>"]
readme = "README.md"
packages = [{include = "upload_service"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "^2.3.3"
flask-oauthlib = "^0.9.6"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
upload-service = "upload_service:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
